remove duplicate class and code
