+++ app/src/androidTest/java/org/eyeseetea/malariacare/test/utils/AssetsFileReader.java
@@ -14,28 +14,6 @@ import java.nio.charset.StandardCharsets;
 
 public class AssetsFileReader {
 
-    public String getStringFromFile(Class clazz, String filename) throws IOException {
-        return getContentFromFile(clazz, filename);
-    }
-
-    @NonNull
-    private String getContentFromFile(Class clazz, String filename) throws IOException {
-        FileInputStream inputStream = new FileInputStream(getFile(clazz, filename));
-        InputStreamReader isr = new InputStreamReader(inputStream);
-        BufferedReader bufferedReader = new BufferedReader(isr);
-        StringBuilder sb = new StringBuilder();
-        String line;
-        while ((line = bufferedReader.readLine()) != null) {
-            sb.append(line);
-        }
-        return sb.toString();
-    }
-
-    private File getFile(Class clazz, String filename) {
-        ClassLoader classLoader = clazz.getClassLoader();
-        URL resource = classLoader.getResource(filename);
-        return new File(resource.getPath());
-    }
 
     public String getStringFromFile(String filename, Context testContext) throws IOException {
